+++ android/src/newarch/com/braze/reactbridge/BrazeReactBridge.kt
@@ -9,59 +9,59 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
         return BrazeReactBridgeImpl.NAME
     }
 
-    override fun getInitialURL(callback: Callback?) {
+    override fun getInitialURL(callback: Callback) {
         // iOS only
     }
 
-    override fun getDeviceId(callback: Callback?) {
+    override fun getDeviceId(callback: Callback) {
         return brazeImpl.getDeviceId(callback)
     }
 
-    override fun changeUser(userId: String?, signature: String?) {
+    override fun changeUser(userId: String, signature: String?) {
         brazeImpl.changeUser(userId, signature)
     }
 
-    override fun getUserId(callback: Callback?) {
+    override fun getUserId(callback: Callback) {
         return brazeImpl.getUserId(callback)
     }
-    
-    override fun setSdkAuthenticationSignature(signature: String?) {
+
+    override fun setSdkAuthenticationSignature(signature: String) {
         brazeImpl.setSdkAuthenticationSignature(signature)
     }
 
-    override fun addAlias(aliasName: String?, aliasLabel: String?) {
+    override fun addAlias(aliasName: String, aliasLabel: String) {
         brazeImpl.addAlias(aliasName, aliasLabel)
     }
 
-    override fun setFirstName(firstName: String?) {
+    override fun setFirstName(firstName: String) {
         brazeImpl.setFirstName(firstName)
     }
 
-    override fun setLastName(lastName: String?) {
+    override fun setLastName(lastName: String) {
         brazeImpl.setLastName(lastName)
     }
 
-    override fun setEmail(email: String?) {
+    override fun setEmail(email: String) {
         brazeImpl.setEmail(email)
     }
 
-    override fun setGender(gender: String?, callback: Callback?) {
+    override fun setGender(gender: String, callback: Callback?) {
         brazeImpl.setGender(gender, callback)
     }
 
-    override fun setLanguage(language: String?) {
+    override fun setLanguage(language: String) {
         brazeImpl.setLanguage(language)
     }
 
-    override fun setCountry(country: String?) {
+    override fun setCountry(country: String) {
         brazeImpl.setCountry(country)
     }
 
-    override fun setHomeCity(homeCity: String?) {
+    override fun setHomeCity(homeCity: String) {
         brazeImpl.setHomeCity(homeCity)
     }
 
-    override fun setPhoneNumber(phoneNumber: String?) {
+    override fun setPhoneNumber(phoneNumber: String) {
         brazeImpl.setPhoneNumber(phoneNumber)
     }
 
@@ -69,103 +69,103 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
         brazeImpl.setDateOfBirth(year.toInt(), month.toInt(), day.toInt())
     }
 
-    override fun registerPushToken(token: String?) {
+    override fun registerPushToken(token: String) {
         brazeImpl.registerPushToken(token)
     }
 
-    override fun addToSubscriptionGroup(groupId: String?, callback: Callback?) {
+    override fun addToSubscriptionGroup(groupId: String, callback: Callback?) {
         brazeImpl.addToSubscriptionGroup(groupId, callback)
     }
 
-    override fun removeFromSubscriptionGroup(groupId: String?, callback: Callback?) {
+    override fun removeFromSubscriptionGroup(groupId: String, callback: Callback?) {
         brazeImpl.removeFromSubscriptionGroup(groupId, callback)
     }
 
     override fun setPushNotificationSubscriptionType(
-        notificationSubscriptionType: String?,
+        notificationSubscriptionType: String,
         callback: Callback?
     ) {
         brazeImpl.setPushNotificationSubscriptionType(notificationSubscriptionType, callback)
     }
 
     override fun setEmailNotificationSubscriptionType(
-        notificationSubscriptionType: String?,
+        notificationSubscriptionType: String,
         callback: Callback?
     ) {
         brazeImpl.setEmailNotificationSubscriptionType(notificationSubscriptionType, callback)
     }
 
-    override fun logCustomEvent(eventName: String?, eventProperties: ReadableMap?) {
+    override fun logCustomEvent(eventName: String, eventProperties: ReadableMap?) {
         brazeImpl.logCustomEvent(eventName, eventProperties)
     }
 
     override fun logPurchase(
-        productId: String?,
-        price: String?,
-        currencyCode: String?,
+        productId: String,
+        price: String,
+        currencyCode: String,
         quantity: Double,
         purchaseProperties: ReadableMap?
     ) {
         brazeImpl.logPurchase(productId, price, currencyCode, quantity.toInt(), purchaseProperties)
     }
 
-    override fun setIntCustomUserAttribute(key: String?, value: Double, callback: Callback?) {
+    override fun setIntCustomUserAttribute(key: String, value: Double, callback: Callback?) {
         brazeImpl.setIntCustomUserAttribute(key, value.toInt(), callback)
     }
 
-    override fun setDoubleCustomUserAttribute(key: String?, value: Double, callback: Callback?) {
+    override fun setDoubleCustomUserAttribute(key: String, value: Double, callback: Callback?) {
         brazeImpl.setDoubleCustomUserAttribute(key, value.toFloat(), callback)
     }
 
-    override fun setBoolCustomUserAttribute(key: String?, value: Boolean, callback: Callback?) {
+    override fun setBoolCustomUserAttribute(key: String, value: Boolean, callback: Callback?) {
         brazeImpl.setBoolCustomUserAttribute(key, value, callback)
     }
 
-    override fun setStringCustomUserAttribute(key: String?, value: String?, callback: Callback?) {
+    override fun setStringCustomUserAttribute(key: String, value: String, callback: Callback?) {
         brazeImpl.setStringCustomUserAttribute(key, value, callback)
     }
 
     override fun setCustomUserAttributeArray(
-        key: String?,
-        value: ReadableArray?,
+        key: String,
+        value: ReadableArray,
         callback: Callback?
     ) {
         brazeImpl.setCustomUserAttributeArray(key, value, callback)
     }
 
     override fun setCustomUserAttributeObjectArray(
-        key: String?,
-        value: ReadableArray?,
+        key: String,
+        value: ReadableArray,
         callback: Callback?
     ) {
         brazeImpl.setCustomUserAttributeObjectArray(key, value, callback)
     }
 
-    override fun setDateCustomUserAttribute(key: String?, value: Double, callback: Callback?) {
+    override fun setDateCustomUserAttribute(key: String, value: Double, callback: Callback?) {
         brazeImpl.setDateCustomUserAttribute(key, value.toInt(), callback)
     }
 
-    override fun setCustomUserAttributeObject(key: String?, value: ReadableMap, merge: Boolean, callback: Callback?) {
+    override fun setCustomUserAttributeObject(key: String?, value: ReadableMap?, merge: Boolean, callback: Callback?) {
         brazeImpl.setCustomUserAttributeObject(key, value, merge, callback)
     }
 
-    override fun addToCustomUserAttributeArray(key: String?, value: String?, callback: Callback?) {
+    override fun addToCustomUserAttributeArray(key: String, value: String, callback: Callback?) {
         brazeImpl.addToCustomAttributeArray(key, value, callback)
     }
 
     override fun removeFromCustomUserAttributeArray(
-        key: String?,
-        value: String?,
+        key: String,
+        value: String,
         callback: Callback?
     ) {
         brazeImpl.removeFromCustomAttributeArray(key, value, callback)
     }
 
-    override fun unsetCustomUserAttribute(key: String?, callback: Callback?) {
+    override fun unsetCustomUserAttribute(key: String, callback: Callback?) {
         brazeImpl.unsetCustomUserAttribute(key, callback)
     }
 
-    override fun incrementCustomUserAttribute(key: String?, value: Double, callback: Callback?) {
+    override fun incrementCustomUserAttribute(key: String, value: Double, callback: Callback?) {
         brazeImpl.incrementCustomUserAttribute(key, value.toInt(), callback)
     }
 
@@ -182,15 +182,15 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
         brazeImpl.launchNewsFeed()
     }
 
-    override fun logNewsFeedCardClicked(cardId: String?) {
+    override fun logNewsFeedCardClicked(cardId: String) {
         brazeImpl.logNewsFeedCardClicked(cardId)
     }
 
-    override fun logNewsFeedCardImpression(cardId: String?) {
+    override fun logNewsFeedCardImpression(cardId: String) {
         brazeImpl.logNewsFeedCardImpression(cardId)
     }
 
-    override fun getNewsFeedCards(promise: Promise?) {
+    override fun getNewsFeedCards(promise: Promise) {
         brazeImpl.getNewsFeedCards(promise)
     }
 
@@ -202,35 +202,35 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
         brazeImpl.requestContentCardsRefresh()
     }
 
-    override fun logContentCardClicked(cardId: String?) {
+    override fun logContentCardClicked(cardId: String) {
         brazeImpl.logContentCardClicked(cardId)
     }
 
-    override fun logContentCardDismissed(cardId: String?) {
+    override fun logContentCardDismissed(cardId: String) {
         brazeImpl.logContentCardDismissed(cardId)
     }
 
-    override fun logContentCardImpression(cardId: String?) {
+    override fun logContentCardImpression(cardId: String) {
         brazeImpl.logContentCardImpression(cardId)
     }
 
-    override fun processContentCardClickAction(cardId: String?) {
+    override fun processContentCardClickAction(cardId: String) {
         brazeImpl.processContentCardClickAction(cardId)
     }
 
-    override fun getContentCards(promise: Promise?) {
+    override fun getContentCards(promise: Promise) {
         brazeImpl.getContentCards(promise)
     }
 
-    override fun getCachedContentCards(promise: Promise?) {
+    override fun getCachedContentCards(promise: Promise) {
         brazeImpl.getCachedContentCards(promise)
     }
 
-    override fun getCardCountForCategories(category: String?, callback: Callback?) {
+    override fun getCardCountForCategories(category: String, callback: Callback?) {
         brazeImpl.getCardCountForCategories(category, callback)
     }
 
-    override fun getUnreadCardCountForCategories(category: String?, callback: Callback?) {
+    override fun getUnreadCardCountForCategories(category: String, callback: Callback?) {
         brazeImpl.getUnreadCardCountForCategories(category, callback)
     }
 
@@ -263,7 +263,7 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
     }
 
     override fun setLocationCustomAttribute(
-        key: String?,
+        key: String,
         latitude: Double,
         longitude: Double,
         callback: Callback?
@@ -289,19 +289,19 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
         brazeImpl.hideCurrentInAppMessage()
     }
 
-    override fun logInAppMessageClicked(inAppMessageString: String?) {
+    override fun logInAppMessageClicked(inAppMessageString: String) {
         brazeImpl.logInAppMessageClicked(inAppMessageString)
     }
 
-    override fun logInAppMessageImpression(inAppMessageString: String?) {
+    override fun logInAppMessageImpression(inAppMessageString: String) {
         brazeImpl.logInAppMessageImpression(inAppMessageString)
     }
 
-    override fun logInAppMessageButtonClicked(inAppMessageString: String?, buttonId: Double) {
+    override fun logInAppMessageButtonClicked(inAppMessageString: String, buttonId: Double) {
         brazeImpl.logInAppMessageButtonClicked(inAppMessageString, buttonId.toInt())
     }
 
-    override fun performInAppMessageAction(inAppMessageString: String?, buttonId: Double) {
+    override fun performInAppMessageAction(inAppMessageString: String, buttonId: Double) {
         brazeImpl.performInAppMessageAction(inAppMessageString, buttonId.toInt())
     }
 
@@ -309,23 +309,23 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
         brazeImpl.requestPushPermission(permissionOptions)
     }
 
-    override fun getAllFeatureFlags(promise: Promise?) {
+    override fun getAllFeatureFlags(promise: Promise) {
         brazeImpl.getAllFeatureFlags(promise)
     }
 
-    override fun getFeatureFlag(flagId: String?, promise: Promise?) {
+    override fun getFeatureFlag(flagId: String, promise: Promise) {
         brazeImpl.getFeatureFlag(flagId, promise)
     }
 
-    override fun getFeatureFlagBooleanProperty(flagId: String?, key: String?, promise: Promise?) {
+    override fun getFeatureFlagBooleanProperty(flagId: String, key: String, promise: Promise) {
         brazeImpl.getFeatureFlagBooleanProperty(flagId, key, promise)
     }
 
-    override fun getFeatureFlagStringProperty(flagId: String?, key: String?, promise: Promise?) {
+    override fun getFeatureFlagStringProperty(flagId: String, key: String, promise: Promise) {
         brazeImpl.getFeatureFlagStringProperty(flagId, key, promise)
     }
 
-    override fun getFeatureFlagNumberProperty(flagId: String?, key: String?, promise: Promise?) {
+    override fun getFeatureFlagNumberProperty(flagId: String, key: String, promise: Promise) {
         brazeImpl.getFeatureFlagNumberProperty(flagId, key, promise)
     }
 
@@ -337,18 +337,16 @@ class BrazeReactBridge(reactContext: ReactApplicationContext): NativeBrazeReactM
         brazeImpl.logFeatureFlagImpression(id)
     }
 
-    override fun setAdTrackingEnabled(adTrackingEnabled: Boolean, googleAdvertisingId: String?) {
+    override fun setAdTrackingEnabled(adTrackingEnabled: Boolean, googleAdvertisingId: String) {
         brazeImpl.setAdTrackingEnabled(adTrackingEnabled, googleAdvertisingId)
     }
 
-    override fun updateTrackingPropertyAllowList(allowList: ReadableMap?) {
+    override fun updateTrackingPropertyAllowList(allowList: ReadableMap) {
         // iOS only
     }
 
-    override fun addListener(eventType: String?) {
-        eventType?.let {
-            brazeImpl.addListener(it)
-        }
+    override fun addListener(eventType: String) {
+        brazeImpl.addListener(eventType)
     }
 
     override fun removeListeners(count: Double) {