+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 9.2.0
+
+##### Added
+- Updates the native iOS version bindings [from Braze Swift SDK 8.2.1 to 8.4.0](https://github.com/braze-inc/braze-swift-sdk/compare/8.2.1...8.4.0#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4ed).
+
+##### Fixed
+- Fixes the Android implementation of `Braze.setCustomUserAttribute()` to correctly handle null values.
+  - Thanks @owonie for your contribution!
+
 ## 9.1.0
 
 ##### Added